ok, basic simulator works
has solar gain, radiation, conduction, convection, and control.

produces answers that make sense.

but it short-cycles (20 min period) with a reasonable deadband,
so the rate of heating seems too high, i.e. it is acting like a
low-mass building.  maybe i need to add some mass.

so this seems like a good place to stop for the night.